created a reactive socket interface, pointed rxjava2 to the specific version of rxjava2 to get the tests to work, updated Agrona to 0.4.13
